[RPC] Support timeout in rref._get_type()
This change is mostly needed for the next diff in this stack, where
rref._get_type() is called in the rpc_async/rpc_sync RRef proxy function and
can block indefinitely if there is no timeout. It will also be useful to have a
timeout argument when we publicize this API to keep it consistent with other
RPC APIs.

Differential Revision: [D25897588](https://our.internmc.facebook.com/intern/diff/D25897588/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D25897588/)!

ghstack-source-id: 119789336
Pull Request resolved: https://github.com/pytorch/pytorch/pull/50498