[tools] Update clang-format linux hash
**Summary**
The new version of `clang-format` for linux64 that was uploaded to S3
earlier this week was dynamically linked to fbcode's custom platform.
A new binary has been uploaded that statically links against `libgcc`
and `libstdc++`, which seems to have fixed this issue. Ideally, all
libraries would be statically linked.

**Test Plan**
`clang-format` workflow passes on this PR and output shows that it
successfully downloaded, verified and ran.

ghstack-source-id: 77136b731736953df82b389bfb21ff83c39afdae
Pull Request resolved: https://github.com/pytorch/pytorch/pull/50520